refactor: 프레젠테이션과 애플리케이션 계층 사이에 Port 레이어를 두고, 두 레이어가 요청하고 응답하는 값을 철저히 분리
