+++ src/main/java/subway/line/presentation/LineController.java
@@ -2,9 +2,10 @@ package subway.line.presentation;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import subway.line.domain.section.dto.SectionRequest;
-import subway.line.domain.section.dto.SectionResponse;
-import subway.line.domain.section.dto.SectionSavingRequest;
+import subway.line.presentation.dto.SectionRequest;
+import subway.line.presentation.dto.SectionResponse;
+import subway.line.presentation.dto.SectionSavingRequest;
+import subway.line.port.LineControllerPort;
 import subway.line.presentation.dto.LineRequest;
 import subway.line.presentation.dto.LineResponse;
 import subway.line.presentation.dto.StationDeletingRequest;
@@ -16,56 +17,56 @@ import java.util.List;
 @RequestMapping("/lines")
 public class LineController {
 
-    private final LineServicePort lineServicePort;
+    private final LineControllerPort lineControllerPort;
 
-    public LineController(LineServicePort lineServicePort) {
-        this.lineServicePort = lineServicePort;
+    public LineController(LineControllerPort lineControllerPort) {
+        this.lineControllerPort = lineControllerPort;
     }
 
     @PostMapping
     public ResponseEntity<LineResponse> createLine(@RequestBody LineRequest lineRequest) {
-        LineResponse lineResponse = lineServicePort.saveLine(lineRequest);
+        LineResponse lineResponse = lineControllerPort.saveLine(lineRequest);
         return ResponseEntity.created(URI.create("/lines/" + lineResponse.getId())).body(lineResponse);
     }
 
     @GetMapping
     public ResponseEntity<List<LineResponse>> findAllLines() {
-        return ResponseEntity.ok(lineServicePort.findLineResponses());
+        return ResponseEntity.ok(lineControllerPort.findLineResponses());
     }
 
     @GetMapping("/{id}")
     public ResponseEntity<LineResponse> findLineById(@PathVariable Long id) {
-        return ResponseEntity.ok(lineServicePort.findLineResponseById(id));
+        return ResponseEntity.ok(lineControllerPort.findLineResponseById(id));
     }
 
     @PutMapping("/{id}")
     public ResponseEntity<Void> updateLine(@PathVariable Long id, @RequestBody LineRequest lineRequest) {
-        lineServicePort.updateLine(id, lineRequest);
+        lineControllerPort.updateLine(id, lineRequest);
         return ResponseEntity.ok().build();
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteLine(@PathVariable Long id) {
-        lineServicePort.deleteLineById(id);
+        lineControllerPort.deleteLineById(id);
         return ResponseEntity.noContent().build();
     }
 
     @PostMapping("/{lineId}/section")
     public ResponseEntity<Void> insertSection(@PathVariable long lineId, @RequestBody SectionSavingRequest sectionSavingRequest) {
-        long savedId = lineServicePort.saveSection(lineId, sectionSavingRequest);
+        long savedId = lineControllerPort.saveSection(lineId, sectionSavingRequest);
         return ResponseEntity.created(URI.create(String.format("/lines/%d/%d", lineId, savedId))).build();
     }
 
     @DeleteMapping("/{lineId}/section")
     public ResponseEntity<Void> deleteStation(@PathVariable long lineId, @RequestBody StationDeletingRequest stationRequest) {
-        lineServicePort.deleteStation(lineId, stationRequest.getStationId());
+        lineControllerPort.deleteStation(lineId, stationRequest.getStationId());
         return ResponseEntity.ok().build();
     }
 
     @GetMapping("/section")
     public ResponseEntity<SectionResponse> findShortestPath(@RequestBody SectionRequest sectionRequest) {
-        final var shortestPath = lineServicePort.findShortestPath(sectionRequest.getStartingStationId(), sectionRequest.getDestinationStationId());
-        final var fare = lineServicePort.calculateFare(shortestPath.getShortestDistance());
+        final var shortestPath = lineControllerPort.findShortestPath(sectionRequest.getStartingStationId(), sectionRequest.getDestinationStationId());
+        final var fare = lineControllerPort.calculateFare(shortestPath.getShortestDistance());
         return ResponseEntity.ok().body(SectionResponse.of(shortestPath, fare));
     }
 }
