+++ src/test/java/subway/integration/SectionIntegrationTest.java
@@ -7,9 +7,9 @@
 //import org.junit.jupiter.api.Test;
 //import org.springframework.http.HttpStatus;
 //import org.springframework.http.MediaType;
-//import subway.line.domain.section.dto.SectionRequest;
-//import subway.line.domain.section.dto.SectionResponse;
-//import subway.line.domain.section.dto.SectionSavingRequest;
+//import subway.line.presentation.dto.SectionRequest;
+//import subway.line.presentation.dto.SectionResponse;
+//import subway.line.presentation.dto.SectionSavingRequest;
 //import subway.line.domain.station.dto.StationResponse;
 //import subway.line.presentation.dto.LineRequest;
 //
